Map parsing 80% complete.
Implemented parsing to follow the rules as described in the subject.
The only issue remaining is to handle map parsing where row[x][y] == 0 && row[x + 1][y] == ' '.
This is a potentially valid map because there may be a '1' at row[x + 2][y].
Will need to create and implement logic to handle this case.